@@ -9,21 +9,22 @@ authors = ["ksdfg <ksdfg123@gmail.com>"]
 license = "GPL-3.0"
 
 [tool.poetry.dependencies]
-python = "3.10.8"
+python = "^3.11"
 dnspython = "^2.2.1"
 emoji = "1.6.1"
-mongoengine = "^0.24.2"
-Pillow = "^9.2.0"
-python-decouple = "^3.6"
-python-telegram-bot = "^13.14"
-spongemock = "^0.3.4"
+mongoengine = "^0.27.0"
+Pillow = "^10.1.0"
+python-decouple = "^3.8"
+python-telegram-bot = "13.15"
+spongemock = "^1.0.0"
 zalgo-text = "^0.6"
 hentai = "^3.2.10"
-telegraph = "^2.1.0"
-pydantic = {extras = ["dotenv"], version = "^1.10.2"}
+telegraph = "^2.2.0"
+pydantic = {extras = ["dotenv"], version = "^2.4.2"}
+pydantic-settings = "^2.0.3"
 
 [tool.poetry.group.dev.dependencies]
-black = "^22.8.0"
+black = "^23.10.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
